Split out tryAndCustomize
